+++ gusto_embedded/src/funcs/paymentConfigsGet.ts
@@ -33,7 +33,7 @@ import { Result } from "../types/fp.js";
  * Get a company's payment configs
  *
  * @remarks
- * Get payment speed for the company and fast payment limit (1-day is only applicable to partners that opt in).
+ * Get payment speed configurations for the company and fast payment limit (1-day is only applicable to partners that opt in).
  *
  * scope: `company_payment_configs:read`
  */