+++ gusto_embedded/src/react-query/reportsGetReportsRequestUuid.ts
@@ -38,6 +38,8 @@ export type ReportsGetReportsRequestUuidQueryData =
  * @remarks
  * Get a company's report given the `request_uuid`. The response will include the report request's status and, if complete, the report URL.
  *
+ * Reports containing PHI are inaccessible with `company_reports:read:tier_2_only` data scope
+ *
  * scope: `company_reports:read`
  */
 export function useReportsGetReportsRequestUuid(
@@ -60,6 +62,8 @@ export function useReportsGetReportsRequestUuid(
  *
  * @remarks
  * Get a company's report given the `request_uuid`. The response will include the report request's status and, if complete, the report URL.
+ *
+ * Reports containing PHI are inaccessible with `company_reports:read:tier_2_only` data scope
  *
  * scope: `company_reports:read`
  */