+++ gusto_embedded/RELEASES.md
@@ -289,3 +289,13 @@ Based on:
 - [typescript v0.6.4] gusto_embedded
 ### Releases
 - [NPM v0.6.4] https://www.npmjs.com/package/@gusto/embedded-api/v/0.6.4 - gusto_embedded
+
+## 2025-07-15 17:19:57
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.582.0 (2.657.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.6.5] gusto_embedded
+### Releases
+- [NPM v0.6.5] https://www.npmjs.com/package/@gusto/embedded-api/v/0.6.5 - gusto_embedded